Introduced the possibility of one agent critiquing another.
From TrainingScript.py you can now call CriticSessions.py, which is another version of Arena.py. This version allows you to have one agent critique the decisions of another during play. So far feedback is only printed to the console as a diagnostic, but it could be extended for an actor-critic architecture.

NOTE: This is only fully implemented for PlayerRandom.py, Myrmidon.py, and DeepPeg.py. Other agents will probably need updates to function signatures for throwCribCards and playCard.